examples/jpg_image_on_canvas/requirements.txt
pygubu
Pillow
